services/service-b-python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
grpcio==1.63.0
grpcio-tools==1.63.0
opentelemetry-sdk==1.25.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-grpc==0.44b0
